Moved json config to examples
