+++ Examples/Detectors/TGeoDetector/CMakeLists.txt
@@ -1,8 +1,7 @@
 add_library(
   ActsExamplesDetectorTGeo SHARED
   src/BuildTGeoDetector.cpp
-  src/TGeoDetector.cpp
-  src/TGeoDetectorOptions.cpp)
+  src/TGeoDetector.cpp)
   
 target_include_directories(
   ActsExamplesDetectorTGeo
@@ -12,7 +11,7 @@ target_link_libraries(
   PUBLIC
     ActsCore ActsPluginIdentification ActsPluginDigitization ActsPluginTGeo
     ActsExamplesFramework ActsExamplesDetectorsCommon
-    ActsExamplesDetectorGeneric)
+    ActsExamplesDetectorGeneric ActsExamplesIoJson)
 
 install(
   TARGETS ActsExamplesDetectorTGeo
