+++ Examples/Io/Json/CMakeLists.txt
@@ -4,13 +4,14 @@ add_library(
   src/JsonMaterialWriter.cpp
   src/JsonSurfacesWriter.cpp
   src/JsonOptionsWriter.cpp
-  src/JsonDigitizationConfig.cpp)
+  src/JsonDigitizationConfig.cpp
+  src/JsonTGeoDetectorConfig.cpp)
 target_include_directories(
   ActsExamplesIoJson
   PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
 target_link_libraries(
   ActsExamplesIoJson
-  PUBLIC ActsCore ActsPluginJson ActsExamplesDigitization ActsExamplesFramework Boost::program_options)
+  PUBLIC ActsCore ActsPluginJson ActsPluginTGeo ActsExamplesDigitization ActsExamplesFramework Boost::program_options)
 
 install(
   TARGETS ActsExamplesIoJson
